+++ CHANGES.md
@@ -1,3 +1,7 @@
+# Unreleased
+- [FIX] Another `NoClassDefFoundError: com.squareup.okhttp.Authenticator` for version 2.6.1 if the
+  optional okhttp dependency was not included.
+
 # 2.6.1 (2016-09-15)
 - [FIX] `NoClassDefFoundError: com.squareup.okhttp.Authenticator` for version 2.6.0 if the optional
   okhttp dependency was not included.
